la til enable

+++ 4_accelerometer/accel.h
@@ -1,7 +1,7 @@
 #ifndef ACCEL_H
 #define ACCEL_H
 
-#define ACCEL_ADDR (uint32_t)(0x01D)
+#define ACCEL_ADDR (uint8_t)(0x01D)
 #define ACCEL_DATA_REG		(uint32_t)(0x01)
 #define ACCEL_CTRL_REG_1	(uint32_t)(0x02A)
 #define ACCEL_200_HZ_ON		(uint32_t)(0x11)
